+++ packages/dart/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add `tailFile` method
 - Add `removeLocalePrefix` method
 - Add `isMatchPathname` method
+- Add `isBotAgent` method
 
 ## 1.1.8 (2025-11-13)